src/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
bidict==0.22.0
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.2
et-xmlfile==1.1.0
fernet==1.0.1
Flask==2.1.3
Flask-RESTful==0.3.8
Flask-SocketIO==5.3.0
Flask-SQLAlchemy==2.4.1
h11==0.13.0
idna==3.3
install==1.3.5
itsdangerous==2.1.2
jellyfish==0.6.1
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
prompt-toolkit==3.0.31
pyaes==1.6.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.2.1
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.23
tabulate==0.8.9
toml==0.10.2
tqdm==4.64.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
us==2.0.2
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.8.5
Werkzeug==2.0.0
wsproto==1.1.0
